Cleanup & rearrange some utils
